Moving test helpers to base tests directory
